+++ src/main/java/com/machina/util/math/MathUtil.java
@@ -77,7 +77,7 @@ public class MathUtil {
 				.add(f[2].scale(-3 * Math.pow(t, 3) + 3 * Math.pow(t, 2))).add(f[3].scale(Math.pow(t, 3)));
 	}
 	
-	public static double[] dirichlet(int size, long seed) {
+	public static double[] dirichlet(int size, int seed) {
 		return new Dirichlet(size).nextDistribution(seed);
 	}
 }
